moved main method to its own class
